requirements.txt
streamlit
scikit-learn
mediapipe
opencv-python-headless
Pillow
joblib
pandas
numpy